Add files via upload
Objective - To find the relationship between gender, urban/rural location and malnourishment among children prevailing in the countries.

Data Source: Child malnutrition database

URL: https://platform.who.int/nutrition/malnutrition-database/database-search

Process:

*The data was selected for the year 2022 using filter.
*Analyzed for the following countries - Burundi, Chad, Kenya, Mali, Mauritania, Nepal, Niger, United Republic of Tanzania
*Unwanted data columns are hidden 
*Chart has been inserted for visualization of the data.
*Pivot table has been created to understand the relationship between gender, urban/rural location and malnourishment prevailing in the countries.

Conclusion: We could see that the sex does not contribute to malnutrition in the country for children. Similarly, Urban/rural location does not contribute to the malnutrition for children in the country.

Additional analysis- From the columns wasting, severe wasting, stunting, underweight, we can notice that among the above-mentioned countries Burundi and Niger are more in danger for malnutrition in children.